Add base code to work on for Baidu map
Code was grabbed from its demo page.
Note the api key (ak) is still the original one as for user's generated
still needs approvement from Baidu. But application was sent, and
possibly wait for 3-5 days from date of this commit.

Search on popup UI is cut all out. Information on UI is not modified yet.